# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Octavia API
  description: |
    Octavia is a video/audio translation backend service with two main components:
    - API Gateway (Go): Handles authentication, job management, and billing
    - AI Worker (Python): Processes translation jobs asynchronously

    This API uses session-based authentication with Redis. All protected endpoints require a valid session cookie.
  version: "1.0.0"
  contact:
    name: Octavia Support
    email: support@octavia.example.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.octavia.example.com
    description: Production server
...
paths:
  /api/v1/auth/signup:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Creates a new user account and returns session information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "200":
          description: User successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security: []
  /api/v1/auth/login:
    post:
      tags:
        - Authentication
      summary: Authenticate user
      description: Logs in a user and creates a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Successful login
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security: []
...
components:
  schemas:
    SignupRequest:
      type: object
      required:
        - email
        - password
        - name
      properties:
        email:
          type: string
          format: email
          description: User email address
          example: user@example.com
        password:
          type: string
          minLength: 8
          description: User password
          example: securepassword123
        name:
          type: string
          description: User full name
          example: John Doe
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User email address
          example: user@example.com
        password:
          type: string
          description: User password
          example: securepassword123
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: User ID
          example: 123e4567-e89b-12d3-a456-426614174000
        email:
          type: string
          format: email
          description: User email
          example: user@example.com
        name:
          type: string
          description: User name
          example: John Doe
        credits:
          type: number
          format: float
          description: Current credit balance
          example: 100.0
        session:
          type: object
          properties:
            user_id:
              type: string
              format: uuid
            created_at:
              type: string
              format: date-time
            user_agent:
              type: string
            ip:
              type: string
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: boolean
          description: Error indicator
          example: true
        message:
          type: string
          description: Error message
          example: Invalid credentials
        code:
          type: integer
          description: HTTP status code
          example: 401
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: octavia_session
      description: |
        Session-based authentication using HTTP-only cookie.
        The session ID is stored in Redis with user information.
        Cookie must be secure and HTTP-only.
    serviceAuth:
      type: apiKey
      in: header
      name: X-Service-API-Key
      description: |
        Service-to-service authentication using API key.
        Required for internal service communication.
        Must match the SERVICE_API_KEY environment variable.
security:
  - sessionAuth: []